move main column generation into parse module
